Override dispatch and initial instead
Wrapping initial instead of initialize_request ensures authentication
failures are caught and handled correctly inside DRF.

If the view raised an IntegrityError, we could't log the response,
because the transaction needed to be rolled back.

Wrapping the usual processing in a transaction of its own ensures that
we can log the response every time. To do that, I had to override
.dispatch().